requirements.txt
# Twitter API interaction (updated for v2 endpoints)
tweepy==4.15.0
requests==2.31.0

# OpenAI API integration (updated for new client)
openai>=1.8.0

# Load environment variables
python-dotenv==1.0.0

# Scheduling tasks
APScheduler==3.10.1

# AWS Integration
boto3==1.34.7
botocore==1.34.7

# File handling utilities
filetype==1.2.0

# Web framework for OAuth flow and admin interface
Flask==2.3.3
Werkzeug==2.3.7

# Testing framework
pytest==8.3.4
pytest-mock==3.11.1
pytest-cov==4.1.0

# Production server
gunicorn==20.1.0

# Utilities for AWS deployment
watchtower==3.0.1   # CloudWatch logging integration
blinker==1.6.2      # Required for Flask signals

# Enhanced error handling and retry logic
tenacity==8.2.3     # Advanced retry mechanisms
cachetools==5.3.2   # For caching responses

# Security
cryptography==41.0.7  # Required for HTTPS with Flask